typo

+++ VisionWalker.h
@@ -24,7 +24,7 @@ class VisionWalker
     private:
     pcl::visualization::CloudViewer* viewer;
     static const int INLIER_THRESHOLD = 20000;
-    static const int RANSAC_DISTANCE_THRESHOLD = 0.03;
+    static const double RANSAC_DISTANCE_THRESHOLD = 0.03;
 };
 
 #endif
